# prototype/public/openapi.yaml
openapi: 3.0.3
info:
  title: PhysaFlow — Prototype API
  description: >
    API REST del prototipo **PhysaFlow**: plataforma de publicación interactiva del
    reporte científico *Stranded Capacity Index*. Los endpoints administrativos
    (`/api/reports*`) están protegidos con sesión de **Better Auth** (cookie
    `better-auth.session_token`) y responden `401` sin sesión válida. Las rutas
    públicas (chat, CSV, BibTeX, health) son consumidas por visitantes del sitio.

    Autenticación: primero usa `POST /api/auth/sign-in/email` desde Swagger UI
    (pestaña Authorize) para obtener la cookie de sesión, y luego podrás ejecutar
    los endpoints protegidos.
  version: 0.1.0
  license:
    name: CC BY-SA 4.0
servers:
  - url: http://localhost:3000
    description: Desarrollo local
  - url: https://prototype-s07-26-team-04.vercel.app
    description: Producción (Vercel)
...
paths:
...
  /api/chat:
    post:
      tags: [Asistente IA]
      summary: Chat con el asistente de IA (streaming)
      description: >
        Público. Envía el historial de mensajes y opcionalmente el `reportId`; el
        servidor responde en **streaming** (chunks de texto plano) usando la ficha
        `aiKnowledge` del reporte como contexto. Si no llega `reportId` (o el
        reporte no existe), usa el **reporte publicado más reciente** — el mismo
        que muestra el home.
      operationId: chatAssistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            example:
              messages:
                - role: user
                  content: ¿Cuál es la mediana global de capacidad varada?
              reportId: clx8abc123def456ghi789jkl
              lang: ES
      responses:
        "200":
          description: Stream de texto (SSE-like, chunks de `text/plain`)
          content:
            text/event-stream:
              schema:
                type: string
              example: "El **31,4%** de la capacidad..."
        "400":
          description: messages ausente o vacío
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No hay reporte en la base de datos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error llamando a Gemini
          content:
            text/plain:
              schema:
                type: string
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Cookie de sesión de Better Auth. Obténla con POST /api/auth/sign-in/email.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: No autorizado
        details:
          type: array
          description: Solo en errores de validación Zod del frontmatter.
          items:
            type: object
...
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, model]
        content:
          type: string
    ChatRequest:
      type: object
      properties:
        messages:
          type: array
          description: Historial completo de la conversación (mínimo 1).
          items:
            $ref: "#/components/schemas/ChatMessage"
        reportId:
          type: string
          description: ID del reporte sobre el que se pregunta. Si falta, se usa el publicado más reciente.
        lang:
          type: string
          description: Idioma (ES/EN) — se envía desde el cliente.
          example: ES
...